--- src/main/java/com/magenic/mobog/exercise3app/Exercise3AppApplication.java
@@ -1,13 +1,0 @@
-package com.magenic.mobog.exercise3app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Exercise3AppApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Exercise3AppApplication.class, args);
-	}
-
-}
