--- src/main/java/com/magenic/mobog/exercise3app/exceptions/InvalidRequestException.java
@@ -1,15 +1,0 @@
-package com.magenic.mobog.exercise3app.exceptions;
-
-import org.springframework.http.HttpStatus;
-
-public class InvalidRequestException extends RuntimeException {
-
-    HttpStatus httpStatus;
-
-    public InvalidRequestException(){
-        this.httpStatus = HttpStatus.BAD_REQUEST;
-    }
-    public HttpStatus getHttpStatus() {
-        return httpStatus;
-    }
-}
