--- src/main/java/com/magenic/mobog/exercise3app/requests/AddAuthorRequest.java
@@ -1,14 +1,0 @@
-package com.magenic.mobog.exercise3app.requests;
-
-public class AddAuthorRequest {
-
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
