--- src/main/java/com/magenic/mobog/exercise3app/requests/AddBookRequest.java
@@ -1,15 +1,0 @@
-package com.magenic.mobog.exercise3app.requests;
-
-public class AddBookRequest {
-    private String title;
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-
-}
