+++ src/main/java/com/magenic/mobog/exercise4app/Exercise4AppApplication.java
@@ -1,0 +1,13 @@
+package com.magenic.mobog.exercise4app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Exercise4AppApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Exercise4AppApplication.class, args);
+	}
+
+}
