+++ src/main/java/com/magenic/mobog/exercise4app/entities/Author.java
@@ -1,0 +1,42 @@
+package com.magenic.mobog.exercise4app.entities;
+
+import com.magenic.mobog.exercise4app.entities.audit.BaseEntity;
+
+import javax.persistence.*;
+import java.util.List;
+
+@Entity
+public class Author extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+
+    @OneToMany(cascade = CascadeType.ALL, mappedBy = "author")
+    private List<Book> books;
+
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public List<Book> getBooks() {
+        return books;
+    }
+
+    public void setBooks(List<Book> books) {
+        this.books = books;
+    }
+}
