+++ src/main/java/com/magenic/mobog/exercise4app/exceptions/EntityNotFoundException.java
@@ -1,0 +1,15 @@
+package com.magenic.mobog.exercise4app.exceptions;
+
+import org.springframework.http.HttpStatus;
+
+public class EntityNotFoundException extends RuntimeException{
+
+    HttpStatus httpStatus;
+
+    public EntityNotFoundException(){
+        this.httpStatus = HttpStatus.NOT_FOUND;
+    }
+    public HttpStatus getHttpStatus() {
+        return httpStatus;
+    }
+}
