+++ src/main/java/com/magenic/mobog/exercise4app/exceptions/InvalidRequestException.java
@@ -1,0 +1,15 @@
+package com.magenic.mobog.exercise4app.exceptions;
+
+import org.springframework.http.HttpStatus;
+
+public class InvalidRequestException extends RuntimeException {
+
+    HttpStatus httpStatus;
+
+    public InvalidRequestException(){
+        this.httpStatus = HttpStatus.BAD_REQUEST;
+    }
+    public HttpStatus getHttpStatus() {
+        return httpStatus;
+    }
+}
