+++ src/main/java/com/magenic/mobog/exercise4app/repositories/AuthorRepository.java
@@ -1,0 +1,13 @@
+package com.magenic.mobog.exercise4app.repositories;
+
+import com.magenic.mobog.exercise4app.entities.Author;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface AuthorRepository extends JpaRepository<Author, Long> {
+    List<Author> findAll();
+    Page<Author> findAll(Pageable page);
+}
