+++ src/main/java/com/magenic/mobog/exercise4app/requests/AddAuthorRequest.java
@@ -1,0 +1,14 @@
+package com.magenic.mobog.exercise4app.requests;
+
+public class AddAuthorRequest {
+
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
