+++ src/main/java/com/magenic/mobog/exercise4app/requests/AddBookRequest.java
@@ -1,0 +1,15 @@
+package com.magenic.mobog.exercise4app.requests;
+
+public class AddBookRequest {
+    private String title;
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+
+}
