+++ src/main/resources/application.yml
@@ -1,0 +1,8 @@
+spring:
+    security:
+        user:
+            name: admin
+            password: password
+management:
+    security:
+        enabled: false
