--- src/test/java/com/magenic/mobog/exercise3app/Exercise3AppApplicationTests.java
@@ -1,13 +1,0 @@
-package com.magenic.mobog.exercise3app;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Exercise3AppApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
