+++ src/test/java/com/magenic/mobog/exercise4app/Exercise3AppApplicationTests.java
@@ -1,0 +1,13 @@
+package com.magenic.mobog.exercise4app;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Exercise3AppApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
